Refactor: split helper modules
