+++ gabriel/__init__.py
@@ -1,18 +1,9 @@
 """Core utilities for the Gabriel project."""
 
-from .utils import (
-    add,
-    delete_secret,
-    divide,
-    floordiv,
-    get_secret,
-    modulo,
-    multiply,
-    power,
-    sqrt,
-    store_secret,
-    subtract,
-)
+from __future__ import annotations
+
+from .arithmetic import add, divide, floordiv, modulo, multiply, power, sqrt, subtract
+from .secrets import delete_secret, get_secret, store_secret
 
 __all__ = [
     "add",
