+++ tests/test_setup_script.py
@@ -7,7 +7,6 @@ import stat
 import subprocess  # nosec B404
 from pathlib import Path
 
-
 REPO_ROOT = Path(__file__).resolve().parents[1]
